+++ CHANGELOG.md
@@ -8,9 +8,9 @@
 
 ### Dependencies
 
-- Bump Android SDK from v6.13.1 to v6.14.0 ([#1287](https://github.com/getsentry/sentry-dart/pull/1287))
-  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#6140)
-  - [diff](https://github.com/getsentry/sentry-java/compare/6.13.1...6.14.0)
+- Bump Android SDK from v6.13.1 to v6.15.0 ([#1287](https://github.com/getsentry/sentry-dart/pull/1287), [#1311](https://github.com/getsentry/sentry-dart/pull/1311))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#6150)
+  - [diff](https://github.com/getsentry/sentry-java/compare/6.13.1...6.15.0)
 
 ## 6.21.0