+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-### Breaking Changes
+### Feature
 
 - Support failedRequestTargets for HTTP Client errors ([#1285](https://github.com/getsentry/sentry-dart/pull/1285))
   - Captures errors for the default range `500-599` if `captureFailedRequests` is enabled