+++ CHANGELOG.md
@@ -17,6 +17,10 @@
 
 - Allow sentry user to control resolution of captured Flutter screenshots ([#1288](https://github.com/getsentry/sentry-dart/pull/1288))
 
+### Features
+
+- Sentry Isolate Extension ([#1266](https://github.com/getsentry/sentry-dart/pull/1266))
+
 ## 6.21.0
 
 ### Features